Miran google api branch (#33)
* Deployed Frontend & Backend

* Placering virker

* Placemarker update

* Update placing  markers not working quite  yet

* Placere markør virker, men at fjerne dem igen gør ikke

* Virk nu for satan

* maps directions

* trying to implement routes

* Revert "trying to implement routes"

This reverts commit 441c7d1f2d8c419cf5e039f92c435406bc1a5afc.

* Revert "Revert "trying to implement routes""

This reverts commit 46433103a3a42c70cb15a6720dcd23280a6a4ed4.

* Virker nu

+++ service/TaxaService.cs
@@ -12,7 +12,7 @@ public class TaxaService
         _taxaRepository = taxaRepository;
     }
 
-    public List<TaxiDTO> GetTaxiPrices(int km, int min, int persons)
+    public List<TaxiDTO> GetTaxiPrices(double km, int min, int persons)
     {
         var taxiDTOs = new List<TaxiDTO>();
         var taxiCompanies = _taxaRepository.GetTaxiCompanies();
